requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
discord.py==1.7.3
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.2
multidict==5.1.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
